DATAGEODE-326 - Add additional 'getOrderedStreamOfBeansByType(..)' utility method in SpringUtils.
This methods also returns beans in order based on bean's class type or bean definitions annotated with the @Order annotation in addition to beans implementing the Ordered interface.

The difference between this method and the 'getBeansOfTypeOrdered(..)' method is that this method uses BeanFactory.getBeanProvider(:Class).orderedStream() internally whereas 'getBeansOfTypeOrdered(..)' uses ListableBeanFactory.getBeansOfType(..) with additional magic.